# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Domino's Beauty API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  description: string;
  price: number;
  imageUrl: string;
  category: string;
  skinTone?: string;
  isBestseller: boolean;
  isFeatured: boolean;
  rating?: number;
  reviewCount?: number;
}

export interface Combo {
  id: number;
  name: string;
  description: string;
  imageUrl: string;
  originalPrice?: number;
  comboPrice: number;
  products?: Product[];
}

export interface CartItem {
  id: number;
  /** @nullable */
  productId?: number | null;
  product?: Product;
  /** @nullable */
  comboId?: number | null;
  /** @nullable */
  comboName?: string | null;
  /** @nullable */
  comboPrice?: number | null;
  quantity: number;
}

export interface Cart {
  items: CartItem[];
  total: number;
  itemCount: number;
}

export interface AddToCartRequest {
  productId: number;
  quantity?: number;
}

export interface AddComboToCartRequest {
  comboId: number;
  comboName: string;
  comboPrice: number;
}

export interface Review {
  id: number;
  author: string;
  text: string;
  rating: number;
}

export interface OrderItem {
  name: string;
  price: number;
  quantity: number;
  /** @nullable */
  comboName?: string | null;
}

export interface Order {
  id: number;
  /** @nullable */
  userId?: string | null;
  total: number;
  items: OrderItem[];
  status: string;
  firstName: string;
  lastName: string;
  email: string;
  /** @nullable */
  address?: string | null;
  createdAt: string;
}

export interface PlaceOrderRequest {
  firstName: string;
  lastName: string;
  email: string;
  address?: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
  loyaltyPoints: number;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export interface AiMatchRequest {
  skinTone: string;
  occasion?: string;
}

export type ListProductsParams = {
  category?: string;
  featured?: boolean;
};