# lib/api-zod/src/generated/types/addComboToCartRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Domino's Beauty API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddComboToCartRequest {
  comboId: number;
  comboName: string;
  comboPrice: number;
}