# lib/api-zod/src/generated/types/cartItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Domino's Beauty API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from "./product";

export interface CartItem {
  id: number;
  /** @nullable */
  productId?: number | null;
  product?: Product;
  /** @nullable */
  comboId?: number | null;
  /** @nullable */
  comboName?: string | null;
  /** @nullable */
  comboPrice?: number | null;
  quantity: number;
}